+++ src/plugins/accounting_storage/mysql/as_mysql_qos.c
@@ -599,7 +599,8 @@ extern int as_mysql_add_qos(mysql_conn_t *mysql_conn, uint32_t uid,
 
 		if (debug_flags & DEBUG_FLAG_DB_QOS)
 			DB_DEBUG(mysql_conn->conn, "query\n%s", query);
-		object->id = mysql_db_insert_ret_id(mysql_conn, query);
+		object->id = (uint32_t)mysql_db_insert_ret_id(
+			mysql_conn, query);
 		xfree(query);
 		if (!object->id) {
 			error("Couldn't add qos %s", object->name);
