+++ src/plugins/accounting_storage/mysql/as_mysql_tres.c
@@ -99,7 +99,8 @@ extern int as_mysql_add_tres(mysql_conn_t *mysql_conn,
 
 		if (debug_flags & DEBUG_FLAG_DB_TRES)
 			DB_DEBUG(mysql_conn->conn, "query\n%s", query);
-		object->id = mysql_db_insert_ret_id(mysql_conn, query);
+		object->id = (uint32_t)mysql_db_insert_ret_id(
+			mysql_conn, query);
 		xfree(query);
 		if (!object->id) {
 			error("Couldn't add tres %s%s%s", object->type,
